Updated with working Effects, woot!
git-svn-id: svn+ssh://svn.code.sf.net/p/winampopenalout/code/trunk@70 b0487f5d-104f-0410-9801-bbc901c8d7b4

+++ ConfigFile.cpp
@@ -46,7 +46,7 @@ namespace WinampOpenALOut {
 		
 		float retVal = 0.0f;
 		if(strlen(buf)>0) {
-			retVal = atof(buf);
+			retVal = (float)atof(buf);
 		}
 		return retVal;
 	}
